requirements.txt
boto3
click
colorama
qbertclient
requests
setuptools
ansible==2.9.9
pytest
kubernetes
testinfra
flask-restx
Werkzeug==0.16.1
mongoengine==0.22.1
mongomock==3.22.1
Flask-Shell2HTTP==1.6.0
pymongo==3.11.3
pyyaml==5.4
flake8==3.8.4
ansible-runner==1.4.7
paramiko==2.7.2
gunicorn==20.0.4
markupsafe==2.0.1